refactor(ui): route truth imports through feature facade
